+++ examples/styleguide/src/templates/ComponentPage/components/ComponentPreview/ComponentPreview.js
@@ -5,11 +5,11 @@ import { LiveProvider, LiveEditor, LiveError, LivePreview } from "react-live"
 import * as components from "../../../../../.cache/components"
 
 import "./prism-theme.css"
+import "./editor.css"
 
 const editorStyles = {
   backgroundColor: `#f2f2f2`,
   boxSizing: `border-box`,
-  padding: `16px`,
 }
 
 class ComponentPreview extends React.Component {